Add integration test D3DFMIQ-2295
Former-commit-id: 41a89efdc430582dd5bfb335f0b9db09eb0557af

+++ src/DeltaShell.Plugins.DelftModels.RealTimeControl/RealTimeControlModel.cs
@@ -1741,9 +1741,12 @@ namespace DeltaShell.Plugins.DelftModels.RealTimeControl
                 return;
             }
 
-            if (outputFileFunctionStore == null)
+            if (outputFileFunctionStore == null && !OutputDocuments.Any())
             {
-                Directory.Delete(targetDirectory, true);
+                if (Directory.Exists(targetDirectory))
+                {
+                    Directory.Delete(targetDirectory, true);
+                }
             }
             else
             {
